+++ bokeh/compat/mplexporter/exporter.py
@@ -9,7 +9,7 @@ import io
 from . import utils
 
 import matplotlib
-from matplotlib import transforms, collections
+from matplotlib import transforms
 from matplotlib.backends.backend_agg import FigureCanvasAgg
 
 class Exporter(object):
@@ -43,7 +43,7 @@ class Exporter(object):
         # Calling savefig executes the draw() command, putting elements
         # in the correct place.
         if fig.canvas is None:
-            canvas = FigureCanvasAgg(fig)
+            fig.canvas = FigureCanvasAgg(fig)
         fig.savefig(io.BytesIO(), format='png', dpi=fig.dpi)
         if self.close_mpl:
             import matplotlib.pyplot as plt
