+++ bokeh/compat/mplexporter/renderers/__init__.py
