+++ bokeh/compat/mplexporter/renderers/base.py
