+++ bokeh/compat/mplexporter/renderers/fake_renderer.py
