+++ bokeh/compat/mplexporter/renderers/vega_renderer.py
