+++ bokeh/compat/mplexporter/renderers/vincent_renderer.py
