+++ bokeh/compat/mplexporter/tools.py
