+++ bokeh/compat/mplexporter/utils.py
