+++ setup.py
@@ -541,9 +541,10 @@ setup(
         'bokeh.charts.builder',
         'bokeh.charts.builder.tests',
         'bokeh.charts.tests',
+        'bokeh.compat',
+        'bokeh.compat.mplexporter',
+        'bokeh.compat.mplexporter.renderers',
         'bokeh.crossfilter',
-        'bokeh.mplexporter',
-        'bokeh.mplexporter.renderers',
         'bokeh.sampledata',
         'bokeh.server',
         'bokeh.server.models',
